Migrate Bundle-related helpers to their own helper
