+++ src/index.ts
@@ -1,5 +1,5 @@
 export * as Calculator from './calculation/Calculator';
 export * as ELMHelpers from './helpers/elm/ELMHelpers';
 export * as ELMDependencyHelpers from './helpers/elm/ELMDependencyHelpers';
-export { extractMeasureFromBundle, extractLibrariesFromBundle } from './calculation/ClauseResultsHelpers';
+export * as MeasureBundleHelpers from './helpers/MeasureBundleHelpers';
 export * from './types';
